+++ test/spec/photon.js
